Changes for readr 2.0.0
In previous version of readr the reading functions took literal data as any
filename with a newline in it or vectors of length > 1. In readr 2.0.0 vectors
of length > 1 are now  assumed to correspond to multiple files. Because of this
we now use a more explicit way to represent literal data, by using the `I()`
function around the input.

readr 2.0.0 by default uses lazy reading. When you first call a reading
function the delimiters in the file are located, but the data is not actually
read until it is used in your program. This can provide substantial speed
improvements for reading character data. Particularly for interactive
exploration of only a subset of a full dataset.

However this also means that problematic values are not necessarily seen
immediately, a warning will be issued the first time a problem is encountered,
which may happen after initial reading.

Use `lazy = FALSE` if you want to uncover all problems up front or want to use
eager reading for any other reason.